+++ buttleofx/gui/paramEditor/wrappers/rgbaWrapper.py
@@ -77,7 +77,7 @@ class RGBAWrapper(QtCore.QObject):
 
     paramType = QtCore.Property(unicode, getParamType, notify=changed)
     text = QtCore.Property(unicode, getText, setText, notify=changed)
-    r = QtCore.Property(int, getValueR, setValueR, notify=changed)
-    g = QtCore.Property(int, getValueG, setValueG, notify=changed)
-    b = QtCore.Property(int, getValueB, setValueB, notify=changed)
-    a = QtCore.Property(int, getValueA, setValueA, notify=changed)
+    r = QtCore.Property(float, getValueR, setValueR, notify=changed)
+    g = QtCore.Property(float, getValueG, setValueG, notify=changed)
+    b = QtCore.Property(float, getValueB, setValueB, notify=changed)
+    a = QtCore.Property(float, getValueA, setValueA, notify=changed)
